requirements.txt
requests==2.31.0
ccxt==4.1.76
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
schedule==1.2.0
python-binance>=1.0.16

